Making my compiler happy when I use cmake.

+++ src/core/functions/evolution/ratematrix/DECRateMatrixFunction.cpp
@@ -16,7 +16,7 @@ using namespace RevBayesCore;
 DECRateMatrixFunction::DECRateMatrixFunction(   const TypedDagNode< RbVector<RbVector<double> > > *dr,
                                                 const TypedDagNode< RbVector<double> > *er,
                                                 const TypedDagNode< RbVector<double> > *rs)
-: TypedFunction<RateGenerator>( new RateMatrix_DECRateMatrix( (size_t)(std::pow(2,er->getValue().size()))) ),
+: TypedFunction<RateGenerator>( new RateMatrix_DECRateMatrix( (size_t)(std::pow(2.0,double(er->getValue().size()) ))) ),
     dispersalRates( dr ),
     extirpationRates( er ),
     rangeSize( rs )
